Create Exponents
In this code, I wrote a function that took in two inputs.
The first is the number. The second is the exponent number and then return the calculation.

In this case....If num =3, and exp = 3, the calculation will be (3*3*3) = 27.
I used the Math.pow...One of the issues I had was I forgot the capitalization of the "M" in Math.Pow. Pow stands for Power.
Another issue I had was I initially wrote 3.math.pow(3), which was incorrect.